refactor: split file and media uploader
